+++ src/wfg/native_ui/internal/ui/core/UIElement.java
@@ -1,4 +1,4 @@
-package wfg.native_ui.ui.core;
+package wfg.native_ui.internal.ui.core;
 
 import java.util.List;
 
@@ -7,6 +7,7 @@ import com.fs.starfarer.api.ui.PositionAPI;
 import com.fs.starfarer.api.ui.UIPanelAPI;
 import com.fs.starfarer.api.util.FaderUtil;
 
+import wfg.native_ui.ui.core.UIElementAPI;
 import wfg.native_ui.ui.event.UIEventBus;
 
 public class UIElement implements UIElementAPI {